+++ packages/mwwm/test/test_entities/test_widget_model.dart
@@ -2,5 +2,6 @@ import 'package:mwwm/src/dependencies/wm_dependencies.dart';
 import 'package:mwwm/src/widget_model.dart';
 
 class TestWidgetModel extends WidgetModel {
-  TestWidgetModel(WidgetModelDependencies baseDependencies) : super(baseDependencies);
+  TestWidgetModel(WidgetModelDependencies baseDependencies)
+      : super(baseDependencies);
 }