+++ packages/mwwm/test/widget_state_test.dart
@@ -6,14 +6,17 @@ import 'package:mwwm/mwwm.dart';
 import 'mocks/widget_model_mock.dart';
 
 class TestWidget extends CoreMwwmWidget<WidgetModelMock> {
-  const TestWidget({required WidgetModelBuilder<WidgetModelMock> widgetModelBuilder, Key? key})
+  const TestWidget(
+      {required WidgetModelBuilder<WidgetModelMock> widgetModelBuilder,
+      Key? key})
       : super(
           key: key,
           widgetModelBuilder: widgetModelBuilder,
         );
 
   @override
-  WidgetState<CoreMwwmWidget<WidgetModelMock>, WidgetModelMock> createWidgetState() {
+  WidgetState<CoreMwwmWidget<WidgetModelMock>, WidgetModelMock>
+      createWidgetState() {
     return TestWidgetState();
   }
 }
@@ -31,7 +34,8 @@ void main() {
     late WidgetModelMock widgetModelMock;
     setUp(() {
       widgetModelMock = WidgetModelMock();
-      testWidget = TestWidget(key: UniqueKey(), widgetModelBuilder: (context) => widgetModelMock);
+      testWidget = TestWidget(
+          key: UniqueKey(), widgetModelBuilder: (context) => widgetModelMock);
     });
     testWidgets(
       'onLoad and onBind called on initState',