+++ packages/mwwm/CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## MAJOR
 
+* **Breaking Change:** Removed nullable types where possible
+* **Breaking Change:** Added generic type of WidgetModel to CoreMwwmWidget and generic type of CoreMwwmWidget to WidgetState to easily get attributes of Widget without cast
 * Marked as deprecated `doFuture` and `doFutureHandleError`
-* Removed nullable types where possible
 
 ## 1.1.1 - 2021-05-30